requirements.txt
ipykernel
fastapi
uvicorn 
javalang
python-multipart
sentence-transformers
python-dotenv
faiss-cpu
openai 
groq 
google-generativeai
langgraph 
langchain